Clear stale setup-form validation error after correction
The interactive setup form only ever set error.content on a failed Enter
and never cleared it, so after a user submitted an invalid Repository owner
and then corrected it, the form kept asserting the value was invalid.

Extract the form into src/setup-form.ts so it can be driven by a test
renderer, and dismiss the validation message on any edit/navigation key so
it reflects the current input. Add a rendered regression test.

Fixes #276
